requirements.txt
apache-airflow==3.1.8
apache-airflow-providers-standard
apache-airflow-providers-postgres
apache-airflow-providers-amazon
apache-airflow-providers-fab

psycopg[binary]>=3.2,<4
kafka-python-ng
websockets>=14,<16
pandas>=2,<3
requests>=2,<3
python-dotenv>=1,<2